requirements.txt
pandas
selenium
webdriver-manager
tkinter
pillow
